refactor: re-think packages and access levels
