+++ src/main/java/org/gridsuite/modification/server/modifications/LargeNetworkModificationExecutionService.java
@@ -1,14 +1,15 @@
-/**
- * Copyright (c) 2024, RTE (http://www.rte-france.com)
+/*
+ * Copyright (c) 2025, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.gridsuite.modification.server.service;
+package org.gridsuite.modification.server.modifications;
 
 import jakarta.annotation.PreDestroy;
 import lombok.NonNull;
+import org.gridsuite.modification.server.service.NetworkModificationObserver;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
@@ -21,12 +22,12 @@ import java.util.function.Supplier;
  * @author Slimane Amar <slimane.amar at rte-france.com>
  */
 @Service
-public class LargeNetworkModificationExecutionService {
+class LargeNetworkModificationExecutionService {
 
     private ThreadPoolExecutor executorService;
 
-    public LargeNetworkModificationExecutionService(@Value("${max-large-concurrent-applications}") int maxConcurrentLargeModifications,
-                                                    @NonNull NetworkModificationObserver networkModificationObserver) {
+    LargeNetworkModificationExecutionService(@Value("${max-large-concurrent-applications}") int maxConcurrentLargeModifications,
+                                             @NonNull NetworkModificationObserver networkModificationObserver) {
         executorService = (ThreadPoolExecutor) Executors.newFixedThreadPool(maxConcurrentLargeModifications);
         networkModificationObserver.createThreadPoolMetric(executorService);
     }
