+++ src/main/java/org/gridsuite/modification/server/network/ModificationNetwork.java
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.gridsuite.modification.server.modifications;
+package org.gridsuite.modification.server.network;
 
 import com.powsybl.iidm.network.Network;
 import org.gridsuite.modification.server.entities.ModificationEntity;
