+++ src/main/java/org/gridsuite/modification/server/network/NetworkVariantsListener.java
@@ -1,10 +1,10 @@
 /*
-  Copyright (c) 2022, RTE (http://www.rte-france.com)
-  This Source Code Form is subject to the terms of the Mozilla Public
-  License, v. 2.0. If a copy of the MPL was not distributed with this
-  file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ * Copyright (c) 2025, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.modification.server.service;
+package org.gridsuite.modification.server.network;
 
 import com.powsybl.commons.extensions.Extension;
 import com.powsybl.iidm.network.Identifiable;
@@ -18,13 +18,13 @@ import java.util.UUID;
 /**
  * @author Nicolas Noir <nicolas.noir at rte-france.com>
  */
-public class NetworkVariantsListener implements NetworkListener {
+class NetworkVariantsListener implements NetworkListener {
 
     private final UUID networkUuid;
 
     private final EquipmentInfosService equipmentInfosService;
 
-    public NetworkVariantsListener(UUID networkUuid, EquipmentInfosService equipmentInfosService) {
+    NetworkVariantsListener(UUID networkUuid, EquipmentInfosService equipmentInfosService) {
         this.networkUuid = networkUuid;
         this.equipmentInfosService = equipmentInfosService;
     }
