+++ src/test/java/org/gridsuite/modification/server/VoltageInitReportTest.java
@@ -20,8 +20,8 @@ import org.gridsuite.modification.server.dto.NetworkModificationResult;
 import org.gridsuite.modification.server.dto.NetworkModificationResult.ApplicationStatus;
 import org.gridsuite.modification.server.dto.ReportInfos;
 import org.gridsuite.modification.server.entities.ModificationEntity;
-import org.gridsuite.modification.server.modifications.ModificationNetwork;
-import org.gridsuite.modification.server.modifications.ModificationNetworkService;
+import org.gridsuite.modification.server.network.ModificationNetwork;
+import org.gridsuite.modification.server.network.ModificationNetworkService;
 import org.gridsuite.modification.server.modifications.NetworkModificationApplicator;
 import org.gridsuite.modification.server.repositories.NetworkModificationRepository;
 import org.gridsuite.modification.server.service.ReportService;
